+++ contracts/services/astroport-withdrawer/src/lib.rs
@@ -1,3 +1,5 @@
+pub mod astroport_cw20;
+pub mod astroport_native;
 pub mod contract;
 pub mod error;
 pub mod msg;
